Fix flake in TracerSettings due to using environment variables
